Trace: Fix a clippy error
Fix a clippy error

+++ devices/src/pci/msix.rs
@@ -450,7 +450,7 @@ impl Msix {
 
         let masked: bool = self.is_func_masked(config);
         let enabled: bool = self.is_enabled(config);
-        trace::msix_write_config(self.dev_id.load(Ordering::Relaxed) as u16, masked, enabled);
+        trace::msix_write_config(self.dev_id.load(Ordering::Relaxed), masked, enabled);
 
         let mask_state_changed = !((self.func_masked == masked) && (self.enabled == enabled));
